wip add dex pkg
