+++ common/dex.go
@@ -2,48 +2,50 @@ package common
 
 // names
 const (
-	// ArgoCDDexServerComponent is the name of the Dex server control plane component
-	ArgoCDDexServerComponent = "argocd-dex-server"
+	// DexServerComponent is the name of the Dex server control plane component
+	DexServerComponent = "dex-server"
 
-	// ArgoCDDefaultDexServiceAccountName is the default Service Account name for the Dex server.
-	ArgoCDDefaultDexServiceAccountName = "argocd-dex-server"
+	// DefaultDexServiceAccountName is the default Service Account name for the Dex server.
+	DefaultDexServiceAccountName = "argocd-dex-server"
 
-	// RedisComponentName is the Redis control plane component
-	DexComponent = "dex"
+	// DexComponent is the dex control plane component
+	DexComponent = "dex-server"
 )
 
 // suffix
 const (
-	DexSuffix = "dex"
+	DexSuffix = "dex-server"
 )
 
 // keys
 const (
-	// ArgoCDKeyDexConfig is the key for dex configuration.
-	ArgoCDKeyDexConfig = "dex.config"
+	// KeyDexConfig is the key for dex configuration.
+	KeyDexConfig = "dex.config"
+
+	DexConfigChangedKey = "dex.config.changed"
 )
 
 // defaults
 const (
 
-	// ArgoCDDefaultDexConfig is the default dex configuration.
-	ArgoCDDefaultDexConfig = ""
+	// DefaultDexConfig is the default dex configuration.
+	DefaultDexConfig = ""
 
-	// ArgoCDDefaultDexImage is the Dex container image to use when not specified.
-	ArgoCDDefaultDexImage = "ghcr.io/dexidp/dex"
+	// DefaultDexImage is the Dex container image to use when not specified.
+	DefaultDexImage = "ghcr.io/dexidp/dex"
 
-	// ArgoCDDefaultDexOAuthRedirectPath is the default path to use for the OAuth Redirect URI.
-	ArgoCDDefaultDexOAuthRedirectPath = "/api/dex/callback"
+	// DefaultDexOAuthRedirectPath is the default path to use for the OAuth Redirect URI.
+	DefaultDexOAuthRedirectPath = "/api/dex/callback"
 
-	// ArgoCDDefaultDexGRPCPort is the default GRPC listen port for Dex.
-	ArgoCDDefaultDexGRPCPort = 5557
+	// DefaultDexGRPCPort is the default GRPC listen port for Dex.
+	DefaultDexGRPCPort = 5557
 
-	// ArgoCDDefaultDexHTTPPort is the default HTTP listen port for Dex.
-	ArgoCDDefaultDexHTTPPort = 5556
+	// DefaultDexHTTPPort is the default HTTP listen port for Dex.
+	DefaultDexHTTPPort = 5556
 
-	// ArgoCDDefaultDexMetricsPort is the default Metrics listen port for Dex.
-	ArgoCDDefaultDexMetricsPort = 5558
+	// DefaultDexMetricsPort is the default Metrics listen port for Dex.
+	DefaultDexMetricsPort = 5558
 
-	// ArgoCDDefaultDexVersion is the Dex container image tag to use when not specified.
-	ArgoCDDefaultDexVersion = "sha256:d5f887574312f606c61e7e188cfb11ddb33ff3bf4bd9f06e6b1458efca75f604" // v2.30.3
+	// DefaultDexVersion is the Dex container image tag to use when not specified.
+	DefaultDexVersion = "sha256:d5f887574312f606c61e7e188cfb11ddb33ff3bf4bd9f06e6b1458efca75f604" // v2.30.3
 )
