+++ controllers/argocd/dex_TOBEREMOVED.go
@@ -1,1 +1,514 @@
 package argocd
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os"
+	"reflect"
+	"strings"
+	"time"
+
+	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
+	"github.com/argoproj-labs/argocd-operator/common"
+	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
+	"gopkg.in/yaml.v2"
+	corev1 "k8s.io/api/core/v1"
+	v1 "k8s.io/api/rbac/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+// DexConnector represents an authentication connector for Dex.
+type DexConnector struct {
+	Config map[string]interface{} `yaml:"config,omitempty"`
+	ID     string                 `yaml:"id"`
+	Name   string                 `yaml:"name"`
+	Type   string                 `yaml:"type"`
+}
+
+func policyRuleForDexServer() []v1.PolicyRule {
+
+	return []v1.PolicyRule{
+		{
+			APIGroups: []string{
+				"",
+			},
+			Resources: []string{
+				"secrets",
+				"configmaps",
+			},
+			Verbs: []string{
+				"get",
+				"list",
+				"watch",
+			},
+		},
+	}
+}
+
+// getDexOAuthRedirectURI will return the OAuth redirect URI for the Dex server.
+func (r *ReconcileArgoCD) getDexOAuthRedirectURI(cr *argoproj.ArgoCD) string {
+	uri := r.getArgoServerURI(cr)
+	return uri + common.ArgoCDDefaultDexOAuthRedirectPath
+}
+
+// getDexResources will return the ResourceRequirements for the Dex container.
+func getDexResources(cr *argoproj.ArgoCD) corev1.ResourceRequirements {
+
+	resources := corev1.ResourceRequirements{}
+
+	// Allow override of resource requirements from CR
+	if cr.Spec.SSO != nil && cr.Spec.SSO.Dex != nil && cr.Spec.SSO.Dex.Resources != nil {
+		resources = *cr.Spec.SSO.Dex.Resources
+	}
+
+	return resources
+}
+
+func getDexConfig(cr *argoproj.ArgoCD) string {
+	config := common.ArgoCDDefaultDexConfig
+
+	// Allow override of config from CR
+	if cr.Spec.ExtraConfig["dex.config"] != "" {
+		config = cr.Spec.ExtraConfig["dex.config"]
+	} else if cr.Spec.SSO != nil && cr.Spec.SSO.Dex != nil && len(cr.Spec.SSO.Dex.Config) > 0 {
+		config = cr.Spec.SSO.Dex.Config
+	}
+	return config
+}
+
+// getDexOAuthClientID will return the OAuth client ID for the given ArgoCD.
+func getDexOAuthClientID(cr *argoproj.ArgoCD) string {
+	return fmt.Sprintf("system:serviceaccount:%s:%s", cr.Namespace, fmt.Sprintf("%s-%s", cr.Name, common.ArgoCDDefaultDexServiceAccountName))
+}
+
+// getDexContainerImage will return the container image for the Dex server.
+//
+// There are three possible options for configuring the image, and this is the
+// order of preference.
+//
+// 1. from the Spec, the spec.sso.dex field has an image and version to use for
+// generating an image reference.
+// 2. from the Environment, this looks for the `ARGOCD_DEX_IMAGE` field and uses
+// that if the spec is not configured.
+// 3. the default is configured in common.ArgoCDDefaultDexVersion and
+// common.ArgoCDDefaultDexImage.
+func getDexContainerImage(cr *argoproj.ArgoCD) string {
+	defaultImg, defaultTag := false, false
+
+	img := ""
+	tag := ""
+
+	if cr.Spec.SSO != nil && cr.Spec.SSO.Dex != nil && cr.Spec.SSO.Dex.Image != "" {
+		img = cr.Spec.SSO.Dex.Image
+	}
+
+	if img == "" {
+		img = common.ArgoCDDefaultDexImage
+		defaultImg = true
+	}
+
+	if cr.Spec.SSO != nil && cr.Spec.SSO.Dex != nil && cr.Spec.SSO.Dex.Version != "" {
+		tag = cr.Spec.SSO.Dex.Version
+	}
+
+	if tag == "" {
+		tag = common.ArgoCDDefaultDexVersion
+		defaultTag = true
+	}
+	if e := os.Getenv(common.ArgoCDDexImageEnvName); e != "" && (defaultTag && defaultImg) {
+		return e
+	}
+	return argoutil.CombineImageTag(img, tag)
+}
+
+// reconcileDexServiceAccount will ensure that the Dex ServiceAccount is configured properly for OpenShift OAuth.
+func (r *ReconcileArgoCD) reconcileDexServiceAccount(cr *argoproj.ArgoCD) error {
+	// if openShiftOAuth set to false in `.spec.sso.dex`, no need to configure it
+	if cr.Spec.SSO == nil || cr.Spec.SSO.Dex == nil || !cr.Spec.SSO.Dex.OpenShiftOAuth {
+		return nil // OpenShift OAuth not enabled, move along...
+	}
+
+	log.Info("oauth enabled, configuring dex service account")
+	sa := newServiceAccountWithName(common.ArgoCDDefaultDexServiceAccountName, cr)
+	if err := argoutil.FetchObject(r.Client, cr.Namespace, sa.Name, sa); err != nil {
+		return err
+	}
+
+	// Get the OAuth redirect URI that should be used.
+	uri := r.getDexOAuthRedirectURI(cr)
+	log.Info(fmt.Sprintf("URI: %s", uri))
+
+	// Get the current redirect URI
+	ann := sa.ObjectMeta.Annotations
+	currentURI, found := ann[common.ArgoCDKeyDexOAuthRedirectURI]
+	if found && currentURI == uri {
+		return nil // Redirect URI annotation found and correct, move along...
+	}
+
+	log.Info(fmt.Sprintf("current URI: %s is not correct, should be: %s", currentURI, uri))
+	if len(ann) <= 0 {
+		ann = make(map[string]string)
+	}
+
+	ann[common.ArgoCDKeyDexOAuthRedirectURI] = uri
+	sa.ObjectMeta.Annotations = ann
+
+	return r.Client.Update(context.TODO(), sa)
+}
+
+// reconcileDexService will ensure that the Service for Dex is present.
+func (r *ReconcileArgoCD) reconcileDexService(cr *argoproj.ArgoCD) error {
+	svc := newServiceWithSuffix("dex-server", "dex-server", cr)
+	if argoutil.IsObjectFound(r.Client, cr.Namespace, svc.Name, svc) {
+
+		// dex uninstallation requested
+		if !UseDex(cr) {
+			log.Info("deleting the existing Dex service because dex uninstallation has been requested")
+			return r.Client.Delete(context.TODO(), svc)
+		}
+		return nil
+	}
+
+	// if Dex installation has not been requested, do nothing
+	if !UseDex(cr) {
+		return nil // Dex is disabled, do nothing
+	}
+
+	svc.Spec.Selector = map[string]string{
+		common.ArgoCDKeyName: nameWithSuffix("dex-server", cr),
+	}
+
+	svc.Spec.Ports = []corev1.ServicePort{
+		{
+			Name:       "http",
+			Port:       common.ArgoCDDefaultDexHTTPPort,
+			Protocol:   corev1.ProtocolTCP,
+			TargetPort: intstr.FromInt(common.ArgoCDDefaultDexHTTPPort),
+		}, {
+			Name:       "grpc",
+			Port:       common.ArgoCDDefaultDexGRPCPort,
+			Protocol:   corev1.ProtocolTCP,
+			TargetPort: intstr.FromInt(common.ArgoCDDefaultDexGRPCPort),
+		},
+	}
+
+	if err := controllerutil.SetControllerReference(cr, svc, r.Scheme); err != nil {
+		return err
+	}
+
+	log.Info(fmt.Sprintf("creating service %s for Argo CD instance %s in namespace %s", svc.Name, cr.Name, cr.Namespace))
+	return r.Client.Create(context.TODO(), svc)
+}
+
+// getDexOAuthClientSecret will return the OAuth client secret for the given ArgoCD.
+func (r *ReconcileArgoCD) getDexOAuthClientSecret(cr *argoproj.ArgoCD) (*string, error) {
+	sa := newServiceAccountWithName(common.ArgoCDDefaultDexServiceAccountName, cr)
+	if err := argoutil.FetchObject(r.Client, cr.Namespace, sa.Name, sa); err != nil {
+		return nil, err
+	}
+
+	// Find the token secret
+	var tokenSecret *corev1.ObjectReference
+	for _, saSecret := range sa.Secrets {
+		if strings.Contains(saSecret.Name, "token") {
+			tokenSecret = &saSecret
+			break
+		}
+	}
+
+	if tokenSecret == nil {
+		// This change of creating secret for dex service account,is due to change of reduction of secret-based service account tokens in k8s v1.24 so from k8s v1.24 no default secret for service account is created, but for dex to work we need to provide token of secret used by dex service account as a oauth token, this change helps to achieve it, in long run we should see do dex really requires a secret or it manages to create one using TokenRequest API or may be change how dex is used or configured by operator
+		secret := &corev1.Secret{
+			ObjectMeta: metav1.ObjectMeta{
+				GenerateName: "argocd-dex-server-token-",
+				Namespace:    cr.Namespace,
+				Annotations: map[string]string{
+					corev1.ServiceAccountNameKey: sa.Name,
+				},
+			},
+			Type: corev1.SecretTypeServiceAccountToken,
+		}
+		err := r.Client.Create(context.TODO(), secret)
+		if err != nil {
+			return nil, errors.New("unable to locate and create ServiceAccount token for OAuth client secret")
+		}
+		err = controllerutil.SetControllerReference(cr, secret, r.Scheme)
+		if err != nil {
+			return nil, err
+		}
+		tokenSecret = &corev1.ObjectReference{
+			Name:      secret.Name,
+			Namespace: cr.Namespace,
+		}
+		sa.Secrets = append(sa.Secrets, *tokenSecret)
+		err = r.Client.Update(context.TODO(), sa)
+		if err != nil {
+			return nil, errors.New("failed to add ServiceAccount token for OAuth client secret")
+		}
+	}
+
+	// Fetch the secret to obtain the token
+	secret := argoutil.NewSecretWithName(cr, tokenSecret.Name)
+	if err := argoutil.FetchObject(r.Client, cr.Namespace, secret.Name, secret); err != nil {
+		return nil, err
+	}
+
+	token := string(secret.Data["token"])
+	return &token, nil
+}
+
+// getOpenShiftDexConfig will return the configuration for the Dex server running on OpenShift.
+func (r *ReconcileArgoCD) getOpenShiftDexConfig(cr *argoproj.ArgoCD) (string, error) {
+	groups := []string{}
+
+	// Allow override of groups from CR
+	if cr.Spec.SSO != nil && cr.Spec.SSO.Dex != nil && cr.Spec.SSO.Dex.Groups != nil {
+		groups = cr.Spec.SSO.Dex.Groups
+	}
+
+	connector := DexConnector{
+		Type: "openshift",
+		ID:   "openshift",
+		Name: "OpenShift",
+		Config: map[string]interface{}{
+			"issuer":       "https://kubernetes.default.svc", // TODO: Should this be hard-coded?
+			"clientID":     getDexOAuthClientID(cr),
+			"clientSecret": "$oidc.dex.clientSecret",
+			"redirectURI":  r.getDexOAuthRedirectURI(cr),
+			"insecureCA":   true, // TODO: Configure for openshift CA,
+			"groups":       groups,
+		},
+	}
+
+	connectors := make([]DexConnector, 0)
+	connectors = append(connectors, connector)
+
+	dex := make(map[string]interface{})
+	dex["connectors"] = connectors
+
+	// add dex config from the Argo CD CR.
+	if err := addDexConfigFromCR(cr, dex); err != nil {
+		return "", err
+	}
+
+	bytes, err := yaml.Marshal(dex)
+	return string(bytes), err
+}
+
+func addDexConfigFromCR(cr *argoproj.ArgoCD, dex map[string]interface{}) error {
+	dexCfgStr := getDexConfig(cr)
+	if dexCfgStr == "" {
+		return nil
+	}
+
+	dexCfg := make(map[string]interface{})
+	if err := yaml.Unmarshal([]byte(dexCfgStr), dexCfg); err != nil {
+		return err
+	}
+
+	for k, v := range dexCfg {
+		dex[k] = v
+	}
+
+	return nil
+}
+
+// reconcileDexConfiguration will ensure that Dex is configured properly.
+func (r *ReconcileArgoCD) reconcileDexConfiguration(cm *corev1.ConfigMap, cr *argoproj.ArgoCD) error {
+	actual := cm.Data[common.ArgoCDKeyDexConfig]
+	desired := getDexConfig(cr)
+
+	// Append the default OpenShift dex config if the openShiftOAuth is requested through `.spec.sso.dex`.
+	if cr.Spec.SSO != nil && cr.Spec.SSO.Dex != nil && cr.Spec.SSO.Dex.OpenShiftOAuth {
+		cfg, err := r.getOpenShiftDexConfig(cr)
+		if err != nil {
+			return err
+		}
+		desired = cfg
+	}
+
+	if actual != desired {
+		// Update ConfigMap with desired configuration.
+		cm.Data[common.ArgoCDKeyDexConfig] = desired
+		if err := r.Client.Update(context.TODO(), cm); err != nil {
+			return err
+		}
+
+		// Trigger rollout of Dex Deployment to pick up changes.
+		deploy := newDeploymentWithSuffix("dex-server", "dex-server", cr)
+		if !argoutil.IsObjectFound(r.Client, deploy.Namespace, deploy.Name, deploy) {
+			log.Info("unable to locate dex deployment")
+			return nil
+		}
+
+		deploy.Spec.Template.ObjectMeta.Labels["dex.config.changed"] = time.Now().UTC().Format("01022006-150406-MST")
+		return r.Client.Update(context.TODO(), deploy)
+	}
+	return nil
+}
+
+// UseDex determines whether Dex resources should be created and configured or not
+func UseDex(cr *argoproj.ArgoCD) bool {
+	if cr.Spec.SSO != nil {
+		return cr.Spec.SSO.Provider.ToLower() == argoproj.SSOProviderTypeDex
+	}
+
+	return false
+}
+
+// reconcileDexDeployment will ensure the Deployment resource is present for the ArgoCD Dex component.
+func (r *ReconcileArgoCD) reconcileDexDeployment(cr *argoproj.ArgoCD) error {
+	deploy := newDeploymentWithSuffix("dex-server", "dex-server", cr)
+
+	AddSeccompProfileForOpenShift(r.Client, &deploy.Spec.Template.Spec)
+
+	dexEnv := proxyEnvVars()
+	if cr.Spec.SSO != nil && cr.Spec.SSO.Dex != nil {
+		dexEnv = append(dexEnv, cr.Spec.SSO.Dex.Env...)
+	}
+
+	deploy.Spec.Template.Spec.Containers = []corev1.Container{{
+		Command: []string{
+			"/shared/argocd-dex",
+			"rundex",
+		},
+		Image: getDexContainerImage(cr),
+		Name:  "dex",
+		Env:   dexEnv,
+		LivenessProbe: &corev1.Probe{
+			ProbeHandler: corev1.ProbeHandler{
+				HTTPGet: &corev1.HTTPGetAction{
+					Path: "/healthz/live",
+					Port: intstr.FromInt(common.ArgoCDDefaultDexMetricsPort),
+				},
+			},
+			InitialDelaySeconds: 60,
+			PeriodSeconds:       30,
+		},
+		Ports: []corev1.ContainerPort{
+			{
+				ContainerPort: common.ArgoCDDefaultDexHTTPPort,
+				Name:          "http",
+			}, {
+				ContainerPort: common.ArgoCDDefaultDexGRPCPort,
+				Name:          "grpc",
+			}, {
+				ContainerPort: common.ArgoCDDefaultDexMetricsPort,
+				Name:          "metrics",
+			},
+		},
+		Resources: getDexResources(cr),
+		SecurityContext: &corev1.SecurityContext{
+			AllowPrivilegeEscalation: boolPtr(false),
+			Capabilities: &corev1.Capabilities{
+				Drop: []corev1.Capability{
+					"ALL",
+				},
+			},
+			RunAsNonRoot: boolPtr(true),
+		},
+		VolumeMounts: []corev1.VolumeMount{{
+			Name:      "static-files",
+			MountPath: "/shared",
+		}},
+	}}
+
+	deploy.Spec.Template.Spec.InitContainers = []corev1.Container{{
+		Command: []string{
+			"cp",
+			"-n",
+			"/usr/local/bin/argocd",
+			"/shared/argocd-dex",
+		},
+		Env:             proxyEnvVars(),
+		Image:           getArgoContainerImage(cr),
+		ImagePullPolicy: corev1.PullAlways,
+		Name:            "copyutil",
+		Resources:       getDexResources(cr),
+		SecurityContext: &corev1.SecurityContext{
+			AllowPrivilegeEscalation: boolPtr(false),
+			Capabilities: &corev1.Capabilities{
+				Drop: []corev1.Capability{
+					"ALL",
+				},
+			},
+			RunAsNonRoot: boolPtr(true),
+		},
+		VolumeMounts: []corev1.VolumeMount{{
+			Name:      "static-files",
+			MountPath: "/shared",
+		}},
+	}}
+
+	deploy.Spec.Template.Spec.ServiceAccountName = fmt.Sprintf("%s-%s", cr.Name, common.ArgoCDDefaultDexServiceAccountName)
+	deploy.Spec.Template.Spec.Volumes = []corev1.Volume{{
+		Name: "static-files",
+		VolumeSource: corev1.VolumeSource{
+			EmptyDir: &corev1.EmptyDirVolumeSource{},
+		},
+	}}
+
+	existing := newDeploymentWithSuffix("dex-server", "dex-server", cr)
+	if argoutil.IsObjectFound(r.Client, cr.Namespace, existing.Name, existing) {
+
+		// dex uninstallation requested
+		if !UseDex(cr) {
+			log.Info("deleting the existing dex deployment because dex uninstallation has been requested")
+			return r.Client.Delete(context.TODO(), existing)
+		}
+		changed := false
+
+		actualImage := existing.Spec.Template.Spec.Containers[0].Image
+		desiredImage := getDexContainerImage(cr)
+		if actualImage != desiredImage {
+			existing.Spec.Template.Spec.Containers[0].Image = desiredImage
+			existing.Spec.Template.ObjectMeta.Labels["image.upgraded"] = time.Now().UTC().Format("01022006-150406-MST")
+			changed = true
+		}
+
+		actualImage = existing.Spec.Template.Spec.InitContainers[0].Image
+		desiredImage = getArgoContainerImage(cr)
+		if actualImage != desiredImage {
+			existing.Spec.Template.Spec.InitContainers[0].Image = desiredImage
+			existing.Spec.Template.ObjectMeta.Labels["image.upgraded"] = time.Now().UTC().Format("01022006-150406-MST")
+			changed = true
+		}
+		updateNodePlacement(existing, deploy, &changed)
+		if !reflect.DeepEqual(existing.Spec.Template.Spec.Containers[0].Env,
+			deploy.Spec.Template.Spec.Containers[0].Env) {
+			existing.Spec.Template.Spec.Containers[0].Env = deploy.Spec.Template.Spec.Containers[0].Env
+			changed = true
+		}
+
+		if !reflect.DeepEqual(existing.Spec.Template.Spec.InitContainers[0].Env,
+			deploy.Spec.Template.Spec.InitContainers[0].Env) {
+			existing.Spec.Template.Spec.InitContainers[0].Env = deploy.Spec.Template.Spec.InitContainers[0].Env
+			changed = true
+		}
+
+		if !reflect.DeepEqual(deploy.Spec.Template.Spec.Containers[0].Resources, existing.Spec.Template.Spec.Containers[0].Resources) {
+			existing.Spec.Template.Spec.Containers[0].Resources = deploy.Spec.Template.Spec.Containers[0].Resources
+			changed = true
+		}
+
+		if changed {
+			return r.Client.Update(context.TODO(), existing)
+		}
+		return nil // Deployment found with nothing to do, move along...
+	}
+
+	// if Dex installation has not been requested, do nothing
+	if !UseDex(cr) {
+		return nil
+	}
+
+	if err := controllerutil.SetControllerReference(cr, deploy, r.Scheme); err != nil {
+		return err
+	}
+
+	log.Info(fmt.Sprintf("creating deployment %s for Argo CD instance %s in namespace %s", deploy.Name, cr.Name, cr.Namespace))
+	return r.Client.Create(context.TODO(), deploy)
+}
