+++ controllers/argocd/sso/dex/dex.go
@@ -4,7 +4,7 @@ import (
 	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
 	"github.com/argoproj-labs/argocd-operator/common"
 	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
-	"github.com/go-logr/logr"
+	"github.com/argoproj-labs/argocd-operator/pkg/util"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -13,7 +13,9 @@ type DexReconciler struct {
 	Client   client.Client
 	Scheme   *runtime.Scheme
 	Instance *argoproj.ArgoCD
-	Logger   logr.Logger
+	Logger   util.Logger
+
+	Server ServerController
 }
 
 var (
@@ -28,6 +30,6 @@ func (dr *DexReconciler) Reconcile() error {
 }
 
 func (dr *DexReconciler) varSetter() {
-	component = common.DexComponent
+	component = common.DexServerComponent
 	resourceName = argoutil.GenerateResourceName(dr.Instance.Name, common.DexSuffix)
 }
