+++ controllers/argocd/sso/dex/interfaces.go
@@ -1,0 +1,5 @@
+package dex
+
+type ServerController interface {
+	GetURI() string
+}
