+++ controllers/argocd/sso/sso.go
@@ -1,22 +1,32 @@
 package sso
 
 import (
-	"github.com/go-logr/logr"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
+	"github.com/argoproj-labs/argocd-operator/controllers/argocd/sso/dex"
+	"github.com/argoproj-labs/argocd-operator/pkg/util"
 )
 
 type SSOReconciler struct {
 	Client   client.Client
 	Scheme   *runtime.Scheme
 	Instance *argoproj.ArgoCD
-	Logger   logr.Logger
+	Logger   util.Logger
+
+	DexController *dex.DexReconciler
 }
 
 func (sr *SSOReconciler) Reconcile() error {
 
 	// controller logic goes here
 	return nil
+}
+
+func (sr *SSOReconciler) GetProvider() argoproj.SSOProviderType {
+	if sr.Instance.Spec.SSO == nil {
+		return ""
+	}
+	return sr.Instance.Spec.SSO.Provider
 }
